Group code analysis extensions
