+++ Funcky.Analyzers/Funcky.Analyzers.CodeFixes/CodeAnalysisExtensions/SyntaxGeneratorExtensions.cs
@@ -1,7 +1,7 @@
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Editing;
 
-namespace Funcky.Analyzers;
+namespace Funcky.Analyzers.CodeAnalysisExtensions;
 
 internal static class SyntaxGeneratorExtensions
 {
