+++ Funcky.Analyzers/Funcky.Analyzers.CodeFixes/Funcky.Analyzers.CodeFixes.targets
@@ -17,8 +17,8 @@
 		<AnalyzerLanguage>cs</AnalyzerLanguage>
 	</PropertyGroup>
 	<ItemGroup>
-		<Compile Include="..\Funcky.Analyzers\SymbolEqualityFunctions.cs" />
-		<Using Include="Funcky.Analyzers.SymbolEqualityFunctions" Static="true" />
+		<Compile Include="..\Funcky.Analyzers\CodeAnalysisExtensions\SymbolEqualityFunctions.cs" Link="CodeAnalysisExtensions\%(Filename)%(Extension)" />
+		<Using Include="Funcky.Analyzers.CodeAnalysisExtensions.SymbolEqualityFunctions" Static="true" />
 	</ItemGroup>
 	<ItemGroup>
 		<PackageReference Include="PolySharp" PrivateAssets="all" />
