+++ Funcky.Analyzers/Funcky.Analyzers/Funcky.Analyzers.targets
@@ -26,7 +26,7 @@
 		<PackageReference Include="Microsoft.CodeAnalysis.CSharp" VersionOverride="$(AnalyzerRoslynVersion)" PrivateAssets="all" />
 	</ItemGroup>
 	<ItemGroup>
-		<Using Include="Funcky.Analyzers.SymbolEqualityFunctions" Static="true" />
+		<Using Include="Funcky.Analyzers.CodeAnalysisExtensions.SymbolEqualityFunctions" Static="true" />
 	</ItemGroup>
 	<ItemGroup>
 		<Compile Update="Resources.Designer.cs">
