+++ Funcky.Analyzers/Funcky.Analyzers/UseWithArgumentNamesAnalyzer.cs
@@ -1,4 +1,5 @@
 using System.Collections.Immutable;
+using Funcky.Analyzers.CodeAnalysisExtensions;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 using Microsoft.CodeAnalysis.Diagnostics;
